+++ src/r_defs.h
@@ -217,12 +217,13 @@ typedef struct line_s
   vertex_t *v1, *v2;     // Vertices, from v1 to v2.
   fixed_t dx, dy;        // Precalculated v2 - v1 for side checking.
   // [FG] extended nodes
-  unsigned short flags;           // Animation related.
-  short special;         
-  short id;
-  int args[5];
-  // [FG] extended nodes
-  unsigned short sidenum[2];      // Visual appearance: SideDefs.
+  // UDMF -- further extend to 32bit
+  uint flags;            // Animation related.
+  int special;           // Special action
+  int id;                // Tag -> id/arg0 split
+  int args[5];           // Hexen-style parameterized actions
+  uint sidenum[2];       // Visual appearance: SideDefs.
+
   fixed_t bbox[4];       // A bounding box, for the linedef's extent
   slopetype_t slopetype; // To aid move clipping.
   sector_t *frontsector; // Front and back sector.
